pyproject.toml
[project]
name = "yt-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp>=2024.0.0",
    "pandas>=2.0.0",
    "pyqt6>=6.6.0",
    "faster-whisper>=1.2.1",
    "static-ffmpeg>=2.13",
    "ffmpeg-python>=0.2.0",
]
